+++ test/lib/package.spec.js
@@ -2,7 +2,7 @@
 
 const { normalize } = require("path");
 
-const { getPackages, getLastPackage } = require("../../lib/package");
+const { getPackages, getLastPackage, getLastPackageSegment } = require("../../lib/package");
 
 describe("lib/package", () => {
   describe("getPackages", () => {
@@ -46,7 +46,6 @@ describe("lib/package", () => {
       expect(getLastPackage(normalize("path/to/node_modules/@scope"))).to.eql(null);
     });
 
-
     it("handles single modules", () => {
       expect(getLastPackage(normalize("node_modules/bar"))).to.eql("bar");
       expect(getLastPackage(normalize("path/to/node_modules/@scope/foo"))).to.eql("@scope/foo");
@@ -59,6 +58,39 @@ describe("lib/package", () => {
         "path/to/node_modules/@scope/foo/node_modules/three/node_modules/four/node_modules"
       )))
         .to.eql("four");
+    });
+  });
+
+  describe("getLastPackageSegment", () => {
+    it("handles empty input", () => {
+      expect(getLastPackageSegment()).to.eql(null);
+      expect(getLastPackageSegment(null)).to.eql(null);
+    });
+
+    it("handles no modules", () => {
+      expect(getLastPackageSegment(normalize(""))).to.eql(null);
+      expect(getLastPackageSegment(normalize("src"))).to.eql(null);
+      expect(getLastPackageSegment(normalize("src/nested/path.js"))).to.eql(null);
+      expect(getLastPackageSegment(normalize("path/to/node_modules"))).to.eql(null);
+      expect(getLastPackageSegment(normalize("path/to/node_modules/@scope"))).to.eql(null);
+    });
+
+    it("handles single modules", () => {
+      expect(getLastPackageSegment(normalize("node_modules/bar"))).to.eql("bar");
+      expect(getLastPackageSegment(normalize("node_modules/bar/src/foo.js")))
+        .to.eql(normalize("bar/src/foo.js"));
+      expect(getLastPackageSegment(normalize("path/to/node_modules/@scope/foo")))
+        .to.eql(normalize("@scope/foo"));
+      expect(getLastPackageSegment(normalize("path/to/node_modules/@scope/foo/index.js")))
+        .to.eql(normalize("@scope/foo/index.js"));
+    });
+
+    it("handles nested modules", () => {
+      expect(getLastPackageSegment(normalize("node_modules/bar/node_modules/@scope/nested-bar")))
+        .to.eql(normalize("@scope/nested-bar"));
+      expect(getLastPackageSegment(
+        normalize("node_modules/bar/node_modules/@scope/nested-bar/path/to/nested.js")
+      )).to.eql(normalize("@scope/nested-bar/path/to/nested.js"));
     });
   });
 });